requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
twilio==9.3.0
python-dotenv==1.0.1
pymongo==4.8.0
numpy==1.26.4
requests==2.32.3
langgraph==0.2.50
